+++ uefi-raw/src/protocol/mod.rs
@@ -18,3 +18,4 @@ pub mod network;
 pub mod rng;
 pub mod shell_params;
 pub mod string;
+pub mod tcg;